20200924 dai
差用户查询历史提交信息

+++ code/pro/src/javaBean/UserInfo.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public class UserInfo {
     private int id;
-    private int number;
+    private long number;
     private List<GameInfo> infoList;
     private String createTime;
 
@@ -17,11 +17,11 @@ public class UserInfo {
         this.id = id;
     }
 
-    public int getNumber() {
+    public long getNumber() {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(long number) {
         this.number = number;
     }
